Add files via upload
Replace this line of code with sys.exit() to make the Quit button function properly. I was testing the Quit button on my M1 Macbook, and I had to force quit to shut the program down. It must freeze up when you press the Quit button. This should make it easier to exit the program.